+++ HISTORY.md
@@ -1,5 +1,6 @@
 unreleased
 ==========
+  * Use `res.headersSent` when available
   * Add the enforceEncoding option for requests without `Accept-Encoding` header
 
 1.7.5 / 2024-10-31